backend/requirements.txt
flask_sqlalchemy
sqlalchemy
sqlalchemy_serializer
flask_restful